% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadShape.R
\name{create_dictionary}
\alias{create_dictionary}
\title{Perform load shape clusing and return cluster centers}
\usage{
create_dictionary(sdata, target.size = 1000, mode = 1, d.metric = 1,
  ths = 0.2, iter.max = 100, nstart = 1, two.step.compress = F,
  verbose = F)
}
\arguments{
\item{sdata}{source data, assume it's already standardized (cleansed and n by p matrix format)}

\item{target.size}{target size of the dictionary (i.e. nubmer of clusters)}

\item{mode}{1: use ths1, 2: use ths2, 3: use ths3, 4: use ths4}

\item{d.metric}{1: use euclidean distance metric, otherwise use cosine distance metric}

\item{ths}{will be transferred to akmeans parameter according to mode setting
\code{ths1}: threshold to decide whether to increase k or not: check sum((sample-assigned center)^2) < ths1*sum(assigned center^2)
\code{ths2}: threshold to decide whether to increase k or not: check all components of |sample-assigned center| < ths2
\code{ths3}: threshold to decide whether to increase k or not: check inner product of (sample,assigned center) > ths3 , this is only for cosine distance metric
\code{ths4}: threshold to decide whether to increase k or not: check sum(abs(sample-assigned center)) < ths4}

\item{iter.max}{maximum iteration setting to be used in kmeans}

\item{two.step.compress}{whether to reduce the dictionary only by hierarchical clustering or hier+use top N shapes.
this option gets activated only when the ratio (original dictionary size before compression/target.size) is larger than 10}

\item{verbose}{whether to show log or not}

\item{n.start}{parameter to be transferred to kmeans}
}
\description{
This function takes load shape data in a standardized matrix format and creates a dictionary of resulting cluster centers
}

