% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadShape.R
\name{draw.top.n.shapes}
\alias{draw.top.n.shapes}
\title{plot top n shapes from encoded result}
\usage{
draw.top.n.shapes(en, en.s = 0, dic, n = 4, show.avg = 0)
}
\arguments{
\item{en}{encoded data / en.s: encoded daily sum / dic: dictionary}

\item{n}{number of top load shapes}

\item{show.avg}{whether to show avg consumption per load shape. if >0, en.s should have the same dimension with en}
}
\description{
plots the top n shapes from an encoded result
}

