% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadShape.R
\name{impute}
\alias{impute}
\title{Meter data imputation function}
\usage{
impute(A, uidx = 4:99)
}
\arguments{
\item{A}{input data matrix, one row per day of meter data, at least two rows should be valid}

\item{uidx}{usage column idx, for example, if A is just hourly consumption data (n by 24 matrix), uidx should be 1:24}
}
\description{
Imputation function to fill in the blanks of an array of meter data. It assumes the input rows come from the same smart meter and are ordered by date.
}

