% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadShape.R
\name{quick.norm}
\alias{quick.norm}
\title{normalize raw smart meter data}
\usage{
quick.norm(A, mod = 2)
}
\arguments{
\item{A}{consumption data matrix, each row corresponds to a daily profile}

\item{mod}{mode of operation with 1: Euclidean (divide by the sqrt of the sum of squares) 2: L1 norm (divide by the sum)}
}
\description{
divide each day of consumption, to get rows (days) of normalized consumption
}

