% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadShape.R
\name{raw2encoded}
\alias{raw2encoded}
\title{Cluster meter data into the best fit shape clusters and return clsuter assignments}
\usage{
raw2encoded(rawdata, is.clean = T, use.all = T, s.size = 1e+05,
  target.size = 1000, mode = 1, d.metric = 1, ths = 0.2,
  iter.max = 100, nstart = 1, two.step.compress = F, verbose = F)
}
\arguments{
\item{is.clean}{whether to do data interpolation or not.
Note that the default interpolation method is very memory and CPU intensive.
You may be better off doing your own interpolation or passing only complete.cases().}

\item{use.all}{whether to use all data to generate a dictionary or not}

\item{s.size}{sample size to use to generate a dictionary}

\item{target.size}{target size of the dictionary (i.e. nubmer of clusters)}

\item{mode}{1: use ths1, 2: use ths2, 3: use ths3, 4: use ths4}

\item{d.metric}{1: use euclidean distance metric, otherwise use cosine distance metric}

\item{ths}{will be transferred to akmeans parameter according to mode setting
\code{ths1}: threshold to decide whether to increase k or not: check sum((sample-assigned center)^2) < ths1*sum(assigned center^2)
\code{ths2}: threshold to decide whether to increase k or not: check all components of |sample-assigned center| < ths2
\code{ths3}: threshold to decide whether to increase k or not: check inner product of (sample,assigned center) > ths3 , this is only for cosine distance metric
\code{ths4}: threshold to decide whether to increase k or not: check sum(abs(sample-assigned center)) < ths4}

\item{iter.max}{maximum iteration setting to be used in kmeans}

\item{two.step.compress}{whether to reduce the dictionary only by hierarchical clustering or hier+use top N shapes.
this option gets activated only when the ratio (original dictionary size before compression/target.size) is larger than 10}

\item{verbose}{whether to show log or not}

\item{rdata}{rawdata of format n by p matrix}

\item{n.start}{parameter to be transferred to kmeans}
}
\description{
Useful for generation of a load shape cluster center dictionary and encoding in one shot
}

