% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadShape.R
\name{reduce.dictionary}
\alias{reduce.dictionary}
\title{reduce the dictionary size hierarchically}
\usage{
reduce.dictionary(dic, cl.size, t.num = 1000, d.metric = 1)
}
\arguments{
\item{dic}{dictionary(=cluster centers) /}

\item{cl.size}{size of each cluster /}

\item{t.num}{target dictionary size}

\item{d.metric}{distance metric to use for similarity comparison == 1: Euclidean, ==2: Cosine}
}
\description{
uses k nearest neighbors (\code{class::knn}) to reduce the size of the passed dictionary to a target number of entries by merging the most similar clusters.
}

