% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadShape.R
\name{shannon.entropy}
\alias{shannon.entropy}
\title{calculate shannon entropy}
\usage{
shannon.entropy(p)
}
\arguments{
\item{p}{is frequency table or probability distribution vector}
}
\description{
a sequence of load shape cluster assignments is a perfect candidate for calculating the Shannon Entropy of a sequence of symbols.
}

