% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadShape.R
\name{shannon.entropy2}
\alias{shannon.entropy2}
\title{calculate shannon entropy}
\usage{
shannon.entropy2(p)
}
\arguments{
\item{p}{a sequence of dictionary encoded load days}
}
\description{
a sequence of load shape cluster assignments is a perfect candidate for calculating the Shannon Entropy of a sequence of symbols.
}

